Remove official support for CUDA 11.0 and 11.1
This is to support using newer CCCL (cuda 11.0 not supported) and simplify the pyflamegpu distribution matrix (11.1).

11.0 is currently builds and passes tests on linux, but does not build on windows.
11.1 currently builds and passes tests on both.

Workarounds and warning specific to these versions are not being removed just incase, and camek will only warn but not error if they are used (as the currently work, just incase 11.2+ is not available somwhere).

Also fixes some typos as and when encountered